# openapi.yaml
openapi: 3.1.0
info:
  title: Pipeline Labs SDK API
  version: 1.0.0
  description: |
    Curated SDK surface for Stainless. This spec is intentionally limited to the
    most relevant programmatic routes for dataset ingestion, AI processing, job
    tracking, downloads, chat, auth verification, billing, and health.
servers:
  - url: https://pipeline-labs.vercel.app
    description: Production
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/datasets:
    get:
      operationId: listDatasets
      tags: [Datasets]
      summary: List datasets
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10, maximum: 100 }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: Paginated datasets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetListResponse'
    delete:
      operationId: deleteDataset
      tags: [Datasets]
      summary: Soft-delete a dataset
      parameters:
        - in: query
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Dataset deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /api/datasets/{id}/preview:
    get:
      operationId: previewDataset
      tags: [Datasets]
      summary: Preview dataset rows and columns
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: rows
          schema: { type: integer, default: 10, maximum: 100 }
      responses:
        '200':
          description: Dataset preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetPreview'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/datasets/{id}/download:
    get:
      operationId: downloadDataset
      tags: [Datasets]
      summary: Download a source or processed dataset
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: format
          schema:
            type: string
            enum: [csv, json, xlsx]
            default: csv
        - in: query
          name: processed
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: File stream
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Dataset:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        original_filename: { type: string }
        file_type: { type: string }
        file_size_bytes: { type: integer }
        row_count: { type: integer }
        column_count: { type: integer }
        status: { type: string }
        storage_url:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
    DatasetListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Dataset'
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
    DatasetPreview:
      type: object
      properties:
        columns:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              type: { type: string }
              sample:
                type: array
                items: { type: string }
        rows:
          type: array
          items:
            type: object
        totalRows: { type: integer }
        totalColumns: { type: integer }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }